// packages/sirius-web/frontend/sirius-web-application/src/views/project-browser/list-projects-area/ListProjectsArea.types.ts
/*******************************************************************************
 * Copyright (c) 2024, 2025 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

import { MRT_RowSelectionState, MRT_SortingState } from 'material-react-table';

export interface ListProjectsAreaProps {}

export interface ListProjectsAreaState {
  pageSize: number;
  startCursor: string | null;
  endCursor: string | null;
  globalFilter: string;
  sorting: MRT_SortingState;
  rowSelection: MRT_RowSelectionState;
}

export interface NoProjectsFoundProps {}

export interface FetchingProjectsProps {}